requirements.txt
asgiref==3.4.1
Django==3.2.6
django-froala-editor==3.2.2
django-restframework==0.0.1
Pillow==8.3.1
pytz==2021.1
sqlparse==0.4.1
